Refatoração de Codigo. corrigido bug ao salvar documento e adicionado funções de alinhamento

+++ GerenciadorDeTarefa.Domain/BlocoDeNotas/ServicoDeGerrenciamentoDeArquivos.cs
@@ -10,19 +10,25 @@ namespace GerenciadorDeTarefa.Domain.BlocoDeNotas
     {
         SaveFileDialog salvarArquivo = new SaveFileDialog();
         OpenFileDialog abrirArquivo = new OpenFileDialog();
-      
 
-        public void SalvarArquivo(string texto)
+
+        public void SalvarArquivo(string texto, string sender)
         {
             try
             {
-                salvarArquivo = new SaveFileDialog { Filter = @"Arquivo txt.(*.txt)|*.txt|Todos os arquivos (*.*)|*.*" };
-                //TODO: Pegar texto do Titulo
-                salvarArquivo.FileName = $"{DateTime.Today:ddMMyyyy}_{DateTime.Now:HHmmss}.txt";
-                if (salvarArquivo.ShowDialog() == DialogResult.OK)
+                var verificaSender = sender.ToString() == "Salvar Como";
+
+                if (!File.Exists(salvarArquivo.FileName)|| verificaSender)
                 {
-                    File.WriteAllText(salvarArquivo.FileName, texto);
+                    salvarArquivo = new SaveFileDialog { Filter = @"Arquivo txt.(*.txt)|*.txt|Todos os arquivos (*.*)|*.*" };
+                    salvarArquivo.FileName = $"{DateTime.Today:ddMMyyyy}_{DateTime.Now:HHmmss}.txt";
+                    if (salvarArquivo.ShowDialog() == DialogResult.OK)
+                    {
+                        File.WriteAllText(salvarArquivo.FileName, texto);
+                    }
+                    return;
                 }
+                File.WriteAllText(salvarArquivo.FileName, texto);
             }
             catch (Exception ex)
             {
@@ -73,18 +79,23 @@ namespace GerenciadorDeTarefa.Domain.BlocoDeNotas
                     if (MessageBox.Show(mensagemAlerta, "Salvar?",
                     MessageBoxButtons.YesNo) == DialogResult.Yes)
                     {
-                        SalvarArquivo(texto);
+                        SalvarArquivo(texto, sender);
                         MessageBox.Show("Salvo com Sucesso");
                     }
                 }
             }
         }
 
-        //Configuração de Fonte
     }
 }
-//TODO: Adicionar verificador ortografico https://www.macoratti.net/vbn_wrde.htm
+//TODO: Adicionar verificador ortografico https://pt.answacode.com/stackoverflow/23834001/como-ativar-a-verificacao-ortografica-em-c-windows-form-application
+//https://docs.microsoft.com/pt-br/dotnet/desktop/wpf/controls/how-to-use-spell-checking-with-a-context-menu?view=netframeworkdesktop-4.8
+//https://www.macoratti.net/vbn_wrde.htm
 //https://social.msdn.microsoft.com/Forums/pt-BR/4c40182d-f4fe-4d1e-b486-0305de8078d9/corretor-ortografico-csharp?forum=vscsharppt
 
 //TODO: usar Save File para salvar formatação do arquivo https://www.youtube.com/watch?v=7er96RwzvRc
 //https://imasters.com.br/back-end/editor-de-texto-com-richtextbox-e-printdocument-parte-01
+
+//Criar tela de historico de notas
+
+//criar função notas fixcadas
